Decreased the stack usage of the FAT filesystem at init
Fixed a bug in the SDFileSystem code. Also changed the way the file handle is allocated on file open so that only one instance of the file handle structure is created instead of three. This reduces the amount of stack memory needed by the open so we no longer nned a stack greater than 2k.